Propagate a possible loadNextBatch error to the consumer.
When loading a batch of data it might be the case that we are loading
a large chunck of data into memory. In this case a CircuitBreakingException
might be thrown to prevent the node from crashing due to OOM.
We'll no propagate this error to the consumer so the query doesn't hang.